fix build issues
