+++ .gitignore
@@ -36,3 +36,6 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 openai-v4.29.0-beta.1.tgz
+.qwen
+data/conversations/
+.idea/
