+++ app/components/ChatSidebar.tsx
@@ -139,7 +139,7 @@ const ChatSidebar = forwardRef<ChatSidebarRef, ChatSidebarProps>(({ onNewChat, o
   const formatTime = (date: Date) => {
     const now = new Date();
     const diffInHours = (now.getTime() - date.getTime()) / (1000 * 60 * 60);
-    
+
     if (diffInHours < 24) {
       return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
     } else {
@@ -255,7 +255,7 @@ const ChatSidebar = forwardRef<ChatSidebarRef, ChatSidebarProps>(({ onNewChat, o
                           <svg className="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24">
                             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
                           </svg>
-                          {formatTime(chat.timestamp)}
+                          <span suppressHydrationWarning>{formatTime(chat.timestamp)}</span>
                         </p>
                       </>
                     )}
