+++ app/function-calling/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import Chat from '@/components/Chat';
+import Chat from '../components/Chat';
 
 const FunctionCalling = () => {
     const [token, setToken] = useState<string | null>(null);
