+++ package.json
@@ -3,13 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build --turbopack",
+    "dev": "next dev",
+    "build": "next build",
     "start": "next start",
     "lint": "eslint"
   },
   "dependencies": {
-    "axios": "^1.12.2",
+    "@fontsource/geist": "^5.2.8",
+    "axios": "^1.13.2",
     "next": "15.5.4",
     "react": "19.1.0",
     "react-dom": "19.1.0"
