+++ vercel.json
@@ -1,0 +1,10 @@
+{
+  "functions": {
+    "app/api/**/*.ts": {
+      "maxDuration": 300
+    }
+  },
+  "buildCommand": "next build",
+  "outputDirectory": ".next",
+  "devCommand": "next dev"
+}
